Move out badge-specific helpers too
