+++ lib/nexus-snapshot-version.js
@@ -1,0 +1,12 @@
+function isNexusSnapshotVersion(version) {
+  var pattern = /(\d+\.)*\d\-SNAPSHOT/;
+  if (version) {
+    return version.match(pattern);
+  } else {
+    return false;
+  }
+}
+
+module.exports = {
+  isNexusSnapshotVersion
+};
